move getNetwork to an api call instead of an action
