+++ api-lib/neynar.ts
@@ -38,8 +38,14 @@ export const fetchUserByUsername = async (username: string) => {
 export const fetchUserByAddress = async (address: string) => {
   try {
     const users = await client.fetchBulkUsersByEthereumAddress([address]);
-    const user = Object.values(users)[0][0];
-    return user;
+    const firstUser = Object.values(users)[0][0];
+    for (const u of Object.values(users)[0]) {
+      // try to find a real one
+      if (u.pfp_url != '' && u.follower_count > 0) {
+        return u;
+      }
+    }
+    return firstUser;
   } catch (err) {
     console.error(
       `Got an error from Neynar attempting fetchBulkUsersByEthereumAddress for address ${address}, prob just 404`,
