+++ hasura/metadata/actions.yaml
@@ -292,17 +292,6 @@ actions:
           value_from_env: HASURA_EVENT_SECRET
     permissions:
       - role: user
-  - name: getNetwork
-    definition:
-      kind: ""
-      handler: '{{HASURA_API_BASE_URL}}/actions/actionManager?action_name=getNetwork'
-      forward_client_headers: true
-      headers:
-        - name: verification_key
-          value_from_env: HASURA_EVENT_SECRET
-    permissions:
-      - role: anon
-      - role: user
   - name: getSimilarProfiles
     definition:
       kind: ""
