+++ hasura/schema/user/schema.graphql
@@ -395,23 +395,6 @@ type MarkClaimedOutput {
   ids: [Int!]!
 }
 
-input NetworkInput {
-  farcaster_id: Int
-  profile_id: Int
-}
-
-type NetworkNode {
-  avatar: String
-  farcaster_id: Int
-  profile_id: Int
-  tier: Int
-  username: String!
-}
-
-type NetworkOutput {
-  nodes: [NetworkNode!]
-}
-
 type OrgMemberResponse {
   OrgMemberResponse: org_members
   id: ID!
@@ -18428,7 +18411,6 @@ type query_root {
   farcaster_accounts_by_pk(profile_id: bigint!): farcaster_accounts
   getGuildInfo(payload: GuildInfoInput!): GuildInfoOutput
   getHeadlines: [HeadlinesOutput!]!
-  getNetwork(payload: NetworkInput!): NetworkOutput
   getSimilarProfiles(payload: SimilarProfileInput!): [SimilarProfileOutput!]!
 
   """
